+++ tests/gd2py/input-output-pairs/class_level_statements.out.py
@@ -21,3 +21,5 @@ def foo():
 	pass
 def bar():
 	pass
+def baz():
+	pass