+++ tests/valid-gd-scripts/variadic_functions.gd
@@ -1,0 +1,10 @@
+func foo(...args):
+	pass
+func bar(a,...args):
+	pass
+func baz(a,...args,):
+	pass
+func bat(a,...args:Array):
+	pass
+func ban(a,...args:Array,):
+	pass